requirements.txt
flask
flask-sqlalchemy
flask-user==1.0.2.2
requests
google-auth
google-auth-oauthlib
google-api-python-client
google-cloud-translate
google-cloud-aiplatform